requirements.txt
aiocontextvars==0.2.2
amqp==5.0.6
asgiref==3.3.4
backcall==0.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
contextvars==2.4
cryptography==3.4.7
decorator==5.0.9
Django==3.2.3
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
dynaconf==3.1.4
gevent==21.1.2
greenlet==1.1.0
hvac==0.10.14
idna==2.10
immutables==0.15
importlib-metadata==4.5.0
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
knox==0.1.11
kombu==5.1.0
loguru==0.5.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.5
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
